Implement `Clone` and `Debug` if the inner types allow it
This introduces support for the `Clone` and `Debug` traits on `TypeMap`,
which allow cloning and printing the map in debug representation,
respectively.

The support is conditional: `Clone` is implemented only if the
underlying storage is `Clone`able. This is accomplished with the
`Clone*Storage` types.

In a similar vein, `Debug` is implemented only if the storage is `Debug`
using the `*DebuggableStorage` types.