Restore remove API that returns removed item (#6)
Restore the original remove API, which returns the removed item (if it
existed), rather than a bool indicating its existence.

This API was adjusted in 466ed27, presumably because its implementation
is more complicated in the presence of debuggable and clonable value
types. But the original API is easily implemented using the same flavor
of unsafe as is used in the `get` method.

Fixes #5.